Move CPUManager ContainerMap to its own package
