+++ pkg/kubelet/cm/cpumanager/BUILD
@@ -3,7 +3,6 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "container_map.go",
         "cpu_assignment.go",
         "cpu_manager.go",
         "fake_cpu_manager.go",
@@ -15,6 +14,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/apis/core/v1/helper/qos:go_default_library",
+        "//pkg/kubelet/cm/cpumanager/containermap:go_default_library",
         "//pkg/kubelet/cm/cpumanager/state:go_default_library",
         "//pkg/kubelet/cm/cpumanager/topology:go_default_library",
         "//pkg/kubelet/cm/cpuset:go_default_library",
@@ -34,7 +34,6 @@ go_library(
 go_test(
     name = "go_default_test",
     srcs = [
-        "container_map_test.go",
         "cpu_assignment_test.go",
         "cpu_manager_test.go",
         "policy_none_test.go",
@@ -69,6 +68,7 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
+        "//pkg/kubelet/cm/cpumanager/containermap:all-srcs",
         "//pkg/kubelet/cm/cpumanager/state:all-srcs",
         "//pkg/kubelet/cm/cpumanager/topology:all-srcs",
     ],
